+++ src/webviews/src/components/EventList.vue
@@ -1,0 +1,8 @@
+<template>
+  <ul
+    class="list-none m-0 p-0 grid gap-x-3 gap-y-3"
+    style="grid-template-columns: max-content max-content minmax(0, 1fr)"
+  >
+    <slot></slot>
+  </ul>
+</template>