+++ src/webviews/src/components/PatchDetail.vue
@@ -7,21 +7,32 @@ import {
 } from '@vscode/webview-ui-toolkit'
 import { ref, computed, toRaw, onMounted, nextTick } from 'vue'
 import { storeToRefs } from 'pinia'
-import Markdown from 'vue3-markdown-it'
-import 'highlight.js/styles/vs2015.css' // TODO: maninak use own style extending this one but using vscode color variables
-import { getIdentityAliasOrId, shortenHash } from 'extensionUtils/string'
+import {
+  getIdentityAliasOrId,
+  shortenHash,
+  truncateMarkdown,
+  maxCharsForUntruncatedMdText,
+  truncateMiddle,
+} from 'extensionUtils/string'
 import { getFormattedDate, getTimeAgo } from 'extensionUtils/time'
 import { notifyExtension } from 'extensionUtils/webview-messaging'
 import { usePatchDetailStore } from '@/stores/patchDetailStore'
-import PatchStatusBadge from './PatchStatusBadge.vue'
-import PatchMajorEvents from './PatchMajorEvents.vue'
-import PatchMetadata from './PatchMetadata.vue'
-import type { Revision } from '../../../types'
+import { scrollToTemplateRef } from '@/utils/scrollToTemplateRef'
+import type { Comment, RadicleIdentity, Revision } from '../../../types'
+import PatchStatusBadge from '@/components/PatchStatusBadge.vue'
+import PatchMajorEvents from '@/components/PatchMajorEvents.vue'
+import PatchMetadata from '@/components/PatchMetadata.vue'
+import Markdown from '@/components/Markdown.vue'
+import Metadatum from '@/components/Metadatum.vue'
+import EventList from '@/components/EventList.vue'
+import EventItem from '@/components/EventItem.vue'
 
 provideVSCodeDesignSystem().register(vsCodeButton(), vsCodeDropdown(), vsCodeOption())
 
-const { patch, authors, firstRevision, latestRevision } = storeToRefs(usePatchDetailStore())
+const { patch, authors, firstRevision, latestRevision, localIdentity, identities } =
+  storeToRefs(usePatchDetailStore())
 
+const revisionSelectorRef = ref<HTMLElement>()
 const revisionOptionsMap = computed(
   () =>
     new Map(
@@ -38,6 +49,27 @@ const selectedRevision = computed(
       ReturnType<(typeof revisionOptionsMap)['value']['get']>
     >,
 )
+
+function selectAndScrollToRevision(revision: Revision) {
+  selectedRevisionOption.value = assembleRevisionOptionLabel(revision)
+  scrollToTemplateRef(revisionSelectorRef.value, {
+    classToAdd: 'pulse-outline',
+    removeAfterMs: 1500,
+  })
+}
+
+function getRevisionHoverTitle(text: string) {
+  return `Click to show Revision details.\n\nRevision description:\n"${text}"`
+}
+
+const commentRefs = ref<InstanceType<typeof EventItem>[]>()
+function scrollToComment(commentId: Comment['id']) {
+  const foundCommentRef = commentRefs.value?.find(
+    (commentEl) => commentEl.$attrs.id === commentId,
+  )
+
+  scrollToTemplateRef(foundCommentRef, { classToAdd: 'pulse-outline', removeAfterMs: 1500 })
+}
 
 function assembleRevisionOptionLabel(revision: Revision): string {
   const id = shortenHash(revision.id)
@@ -69,6 +101,44 @@ const selectedRevisionRejectedReviews = computed(() =>
   selectedRevision.value.reviews.filter((review) => review.verdict === 'reject'),
 )
 
+const patchEvents = computed(() =>
+  [
+    patch.value.revisions.map(
+      (revision) => ({ kind: 'revision', ts: revision.timestamp, revision }) as const,
+    ),
+    patch.value.revisions.flatMap((revision) =>
+      revision.reviews.map(
+        (review) => ({ kind: 'review', ts: review.timestamp, review, revision }) as const,
+      ),
+    ),
+    patch.value.revisions.flatMap((revision) =>
+      revision.discussions.map(
+        (discussion) =>
+          ({ kind: 'discussion', ts: discussion.timestamp, discussion, revision }) as const,
+      ),
+    ),
+    patch.value.merges.map(
+      (merge) =>
+        ({
+          kind: 'merge',
+          ts: merge.timestamp,
+          merge,
+          revision: patch.value.revisions.find((revision) => revision.id === merge.revision),
+        }) as const,
+    ),
+  ]
+    .flat()
+    .sort((ev1, ev2) => ev2.ts - ev1.ts),
+)
+
+// TODO: delete resolveRadicleIdentity when httpd returns a full RadicleIdentity for reactions
+// When doing so, also delete:
+//  - `localIdentity` from `PatchDetailInjectedState`
+//  - `identities` from `patchDetailStore`
+function resolveRadicleIdentity(id: string): RadicleIdentity | undefined {
+  return identities.value.find((identity) => identity.id.includes(id))
+}
+
 function refetchPatchData() {
   notifyExtension({ command: 'refreshPatchData', payload: { patchId: patch.value.id } })
 }
@@ -98,15 +168,13 @@ onMounted(() => {
       langTagEl.classList.add('langTag')
 
       highlightedCodeEl.insertBefore(langTagEl, highlightedCodeEl.firstChild)
-
-      console.info('inserted')
     })
   })
 })
 </script>
 
 <template>
-  <article class="pt-[20px] flex flex-col gap-12">
+  <article class="flex flex-col gap-12">
     <!-- TODO: maninak make h2s (and maybe also header?) sticky -->
     <header class="flex gap-4 justify-between">
       <div class="flex flex-wrap gap-4 items-center">
@@ -156,27 +224,24 @@ onMounted(() => {
       </aside>
     </header>
     <main class="flex flex-col gap-12">
-      <section id="patch">
+      <section>
         <h2 class="text-lg mt-0 mb-3">#&nbsp;Patch</h2>
         <PatchMetadata />
-        <!-- TODO: maninak add button to reveal Patch item in Patches view -->
-        <h1 class="my-4 text-3xl font-mono">
-          <Markdown :source="patch.title" class="parsed-md" />
-        </h1>
-        <!-- TODO: maninak add control to toggle raw/parsed markdown -->
-        <Markdown :source="firstRevision.description" class="parsed-md text-sm" />
+        <h1 class="my-4 text-3xl font-mono"><Markdown :source="patch.title" /></h1>
+        <Markdown :source="firstRevision.description" class="text-sm" />
       </section>
 
-      <section id="revision">
+      <section>
         <h2
-          class="flex items-center gap-[0.5em] text-lg mt-0 flex-wrap mb-3"
+          class="max-w-max flex items-center gap-[0.5em] text-lg mt-0 flex-wrap mb-3"
           title="Select a patch revision"
         >
           <label for="revision-selector" class="cursor-pointer">#&nbsp;Revision</label>
           <vscode-dropdown
             id="revision-selector"
+            ref="revisionSelectorRef"
             v-model="selectedRevisionOption"
-            class="font-mono"
+            class="font-mono rounded-none"
           >
             <vscode-option
               v-for="revisionOption in revisionOptionsMap.keys()"
@@ -186,12 +251,10 @@ onMounted(() => {
             >
           </vscode-dropdown>
         </h2>
-        <div class="*:min-h-[1.5em]">
-          <div class="flex flex-row items-center w-max gap-[0.5em] group">
-            Id:
-            <pre :title="selectedRevision.id">{{ shortenHash(selectedRevision.id) }}</pre>
+        <Metadatum label="Id">
+          <pre :title="selectedRevision.id">{{ shortenHash(selectedRevision.id) }}</pre>
+          <template #aside>
             <vscode-button
-              class="invisible group-hover:visible"
               appearance="icon"
               title="Copy Revision Identifier to Clipboard"
               @click="
@@ -203,161 +266,74 @@ onMounted(() => {
             >
               <span class="codicon codicon-copy"></span>
             </vscode-button>
-          </div>
-          <div class="flex flex-row items-center w-max gap-[0.5em]">
-            Author:
-            <pre :title="selectedRevision.author.id" class="flex gap-[0.5em] w-max">{{
-              getIdentityAliasOrId(selectedRevision.author)
-            }}</pre>
-          </div>
-          <div
-            v-if="
-              selectedRevisionAcceptedReviews.length || selectedRevisionRejectedReviews.length
-            "
-            class="flex flex-row items-center w-max gap-[0.5em]"
-          >
-            Reviews:
-            <div
+          </template>
+        </Metadatum>
+        <Metadatum label="Author">
+          <pre :title="selectedRevision.author.id">{{
+            getIdentityAliasOrId(selectedRevision.author)
+          }}</pre>
+        </Metadatum>
+        <Metadatum
+          v-if="
+            selectedRevisionAcceptedReviews.length || selectedRevisionRejectedReviews.length
+          "
+          label="Reviews"
+        >
+          <span class="flex items-center gap-[1em]">
+            <span
               v-if="selectedRevisionAcceptedReviews.length"
-              class="flex flex-row items-center w-max gap-[1em]"
+              class="flex items-start gap-x-[0.5em]"
             >
-              <div class="flex flex-row items-center w-max gap-[0.5em]">
-                <span class="codicon codicon-thumbsup" title="Accepted revision"></span>
-                <div
-                  v-for="review in selectedRevisionAcceptedReviews"
-                  :key="review.timestamp"
-                  class="flex flex-row items-center w-max gap-[0.5em]"
-                >
-                  <pre :title="selectedRevision.author.id" class="flex gap-[0.5em] w-max">{{
-                    getIdentityAliasOrId(selectedRevision.author)
-                  }}</pre>
-                </div>
-              </div>
-              <div
-                v-if="selectedRevisionRejectedReviews.length"
-                class="flex flex-row items-center w-max gap-[0.5em]"
-              >
-                <span class="codicon codicon-thumbsdown" title="Rejected revision"></span>
-                <div
-                  v-for="review in selectedRevisionRejectedReviews"
-                  :key="review.timestamp"
-                  class="flex flex-row items-center w-max gap-[0.5em]"
-                >
-                  <pre :title="selectedRevision.author.id" class="flex gap-[0.5em] w-max">{{
-                    getIdentityAliasOrId(selectedRevision.author)
-                  }}</pre>
-                </div>
-              </div>
-            </div>
-          </div>
-          <div class="flex flex-row items-center w-max gap-[0.5em]">
-            Date:
-            <pre>{{ getFormattedDate(selectedRevision.timestamp) }}</pre>
-          </div>
-          <div class="flex flex-row items-center w-max gap-[0.5em]">
-            Latest commit:
-            <pre :title="selectedRevision.oid">{{ shortenHash(selectedRevision.oid) }}</pre>
-          </div>
-          <div class="flex flex-row items-center w-max gap-[0.5em]">
-            Based on commit:
-            <pre :title="selectedRevision.base">{{ shortenHash(selectedRevision.base) }}</pre>
-          </div>
-        </div>
+              <span class="codicon codicon-thumbsup" title="Accepted revision"></span>
+              <span v-for="review in selectedRevisionAcceptedReviews" :key="review.timestamp">
+                <pre :title="selectedRevision.author.id">{{
+                  getIdentityAliasOrId(selectedRevision.author)
+                }}</pre>
+              </span>
+            </span>
+            <span
+              v-if="selectedRevisionRejectedReviews.length"
+              class="flex items-start gap-x-[0.5em]"
+            >
+              <span class="codicon codicon-thumbsdown" title="Rejected revision"></span>
+              <span v-for="review in selectedRevisionRejectedReviews" :key="review.timestamp">
+                <pre :title="selectedRevision.author.id">{{
+                  getIdentityAliasOrId(selectedRevision.author)
+                }}</pre>
+              </span>
+            </span>
+          </span>
+        </Metadatum>
+        <Metadatum label="Date">
+          <span
+            :title="new Date(selectedRevision.timestamp * 1000).toISOString()"
+            class="font-mono"
+            >{{ getFormattedDate(selectedRevision.timestamp) }}</span
+          >
+        </Metadatum>
+        <Metadatum label="Latest commit">
+          <pre :title="selectedRevision.oid">{{ shortenHash(selectedRevision.oid) }}</pre>
+        </Metadatum>
+        <Metadatum label="Based on commit">
+          <pre :title="selectedRevision.base">{{ shortenHash(selectedRevision.base) }}</pre>
+        </Metadatum>
         <div v-if="selectedRevision.description" class="mt-4">
           <details v-if="shouldHideRevisionDescription">
-            <summary
-              style="color: var(--vscode-foreground)"
-              title="Click to expand/collapse revision description"
+            <summary style="color: var(--vscode-foreground)" title="Click to expand/collapse"
               >Description</summary
             >
-            <Markdown
-              :source="selectedRevision.description"
-              class="parsed-md text-sm mt-[0.25em]"
-            />
+            <Markdown :source="selectedRevision.description" class="mt-[0.25em] text-sm" />
           </details>
-          <Markdown v-else :source="selectedRevision.description" class="parsed-md text-sm" />
+          <Markdown v-else :source="selectedRevision.description" class="text-sm" />
         </div>
       </section>
 
-      <section id="activity">
+      <section>
+        <!-- TODO: add button to expand/collapse all -->
         <h2 class="text-lg mt-0 mb-3">#&nbsp;Activity</h2>
-        <ul class="timeline -ml-[24px] flex flex-col gap-4">
-          <li class="-ml-4 flex gap-3 list-none items-center">
-            <span
-              title="Radicle patch revision"
-              class="p-[2px] rounded-full no-underline codicon codicon-pulse"
-              style="
-                background-color: color-mix(
-                  in srgb-linear,
-                  var(--vscode-editor-background),
-                  var(--vscode-editor-foreground) 5%
-                );
-              "
-            ></span>
-            <span>
-              Patch created with revision <pre>cdba53e</pre> by
-              <pre>danielkalman 21 hours ago</pre>
-            </span>
-          </li>
-          <li class="-ml-4 flex gap-3 list-none items-center">
-            <span
-              title="Radicle patch revision"
-              class="p-[2px] rounded-full no-underline codicon codicon-pulse"
-              style="
-                background-color: color-mix(
-                  in srgb-linear,
-                  var(--vscode-editor-background),
-                  var(--vscode-editor-foreground) 5%
-                );
-              "
-            ></span>
-            <span>
-              Patch updated with revision <pre>829bee2</pre> by
-              <pre>danielkalman 10 hours ago</pre>
-            </span>
-          </li>
-          <!-- TODO: maninak list committers with gravatar if available and email as tooltip -->
-          <li class="-ml-4 flex gap-3 list-none">
-            <span
-              title="Git commit"
-              class="px-[2px] no-underline before:bg-vscode-editor-background codicon codicon-git-commit"
-            ></span>
-            <div class="flex items-center">
-              <pre>Add embeds to patch descriptions</pre>
-              <vscode-button
-                class="ml-1"
-                appearance="icon"
-                title="Show/hide additional commit info"
-              >
-                <span class="codicon codicon-ellipsis"></span>
-              </vscode-button>
-            </div>
-            <pre class="ml-4" tile="87afc6287afc6287afc62">87afc62</pre>
-            &ndash;
-            <pre title="me@dnlklmn.dev">dnlklmn</pre>
-            <pre title=""></pre> <pre>28 hours ago</pre>
-          </li>
-          <li class="-ml-4 flex gap-3 list-none">
-            <span
-              title="Git commit"
-              class="px-[2px] no-underline before:bg-vscode-editor-background codicon codicon-git-commit"
-            ></span>
-            <div class="flex items-center">
-              <pre>Add <code>--no-announce</code> to patch comment creation for stdout</pre>
-              <vscode-button
-                class="ml-1"
-                appearance="icon"
-                title="Show/hide additional commit info"
-              >
-                <span class="codicon codicon-ellipsis"></span>
-              </vscode-button>
-            </div>
-            <pre class="ml-4" title="cc08490cc08490cc08490">cc08490</pre>
-            &ndash;
-            <pre title="me@sebastinez.dev">Sebastian Martinez</pre>
-            <pre title=""></pre> <pre>11 hours ago</pre>
-          </li>
-          <li class="-ml-4 flex gap-3 items-center list-none">
+        <EventList>
+          <!-- TODO: list committer's email as tooltip -->
+          <!--<div class="grid grid-cols-subgrid gap-x-3 items-center">
             <span
               title="Git commit"
               class="px-[2px] no-underline before:bg-vscode-editor-background codicon codicon-git-commit"
@@ -372,33 +348,227 @@ onMounted(() => {
                 <span class="codicon codicon-ellipsis"></span>
               </vscode-button>
             </div>
-            <pre class="ml-4" title="aed8538aed8538aed8538">aed8538</pre>
+            <pre title="aed8538aed8538aed8538">aed8538</pre>
             &ndash;
             <pre title="me@sebastinez.dev">Sebastian Martinez</pre>
             <pre>10 hours ago</pre>
-          </li>
-        </ul>
+          </div> -->
+          <template v-for="event in patchEvents" :key="event.ts">
+            <EventItem
+              v-if="event.kind === 'revision'"
+              :when="event.ts"
+              codicon="codicon-versions"
+            >
+              {{ event.revision.id === firstRevision.id ? 'Patch and revision' : 'Revision' }}
+              <span
+                @click="selectAndScrollToRevision(event.revision)"
+                :title="getRevisionHoverTitle(event.revision.description)"
+                class="font-mono hover:cursor-pointer"
+                >{{ shortenHash(event.revision.id) }}</span
+              >
+              created by
+              <span :title="event.revision.author.id" class="font-mono">{{
+                getIdentityAliasOrId(event.revision.author)
+              }}</span>
+            </EventItem>
+            <EventItem
+              v-else-if="event.kind === 'review'"
+              :when="event.ts"
+              :codicon="
+                event.review.verdict === 'accept'
+                  ? 'codicon-thumbsup'
+                  : event.review.verdict === 'reject'
+                    ? 'codicon-thumbsdown'
+                    : 'codicon-feedback'
+              "
+            >
+              Review
+              <span v-if="event.review.verdict" class="font-mono">
+                {{ `${event.review.verdict}ing` }}
+              </span>
+              <template v-else>with <span class="font-mono">no verdict</span> for</template>
+              revision
+              <span
+                @click="selectAndScrollToRevision(event.revision)"
+                :title="getRevisionHoverTitle(event.revision.description)"
+                class="font-mono hover:cursor-pointer"
+                >{{ shortenHash(event.revision.id) }}</span
+              >
+              posted
+              <span v-if="event.review.inline?.length">with code-inlined comments</span>
+              by
+              <span :title="event.revision.author.id" class="font-mono">{{
+                getIdentityAliasOrId(event.revision.author)
+              }}</span>
+              <template v-if="event.review.summary">
+                <details v-if="event.review.summary && event.review.comment">
+                  <summary
+                    style="color: var(--vscode-foreground)"
+                    title="Click to expand/collapse"
+                    class="mt-1 max-w-prose break-words text-sm font-mono"
+                    >{{ event.review.summary }}</summary
+                  >
+                  <Markdown :source="event.review.comment" class="mt-[0.25em] text-sm" />
+                </details>
+                <p
+                  v-else-if="event.review.summary && !event.review.comment"
+                  class="mt-1 -mb-[0.2em] max-w-prose break-words text-sm font-mono"
+                  >{{ event.review.summary }}</p
+                >
+              </template>
+            </EventItem>
+            <EventItem
+              v-else-if="event.kind === 'discussion'"
+              ref="commentRefs"
+              :id="event.discussion.id"
+              :when="event.ts"
+              :codicon="
+                event.discussion.resolved ? 'codicon-comment' : 'codicon-comment-unresolved'
+              "
+            >
+              Comment
+              <span v-if="!event.discussion.resolved"
+                >(<span class="font-mono">unresolved</span>)</span
+              >
+              posted on revision
+              <span
+                @click="selectAndScrollToRevision(event.revision)"
+                :title="getRevisionHoverTitle(event.revision.description)"
+                class="font-mono hover:cursor-pointer"
+                >{{ shortenHash(event.revision.id) }}</span
+              >
+              by
+              <span :title="event.discussion.author.id" class="font-mono">{{
+                getIdentityAliasOrId(event.discussion.author)
+              }}</span>
+              <span v-if="event.discussion.replyTo">
+                in reply to
+                <span
+                  @click="scrollToComment(event.discussion.replyTo)"
+                  title="Click to show parent comment"
+                  class="font-mono hover:cursor-pointer"
+                  >another</span
+                ></span
+              >
+              <details
+                v-if="event.discussion.body.length > maxCharsForUntruncatedMdText"
+                class="[&_summary]:open:opacity-50"
+              >
+                <summary
+                  style="color: var(--vscode-foreground)"
+                  title="Click to expand/collapse"
+                  class="mt-1 max-w-prose text-sm font-mono"
+                  >{{ truncateMarkdown(event.discussion.body) }}</summary
+                >
+                <Markdown :source="event.discussion.body" class="mt-[0.25em] text-sm" />
+              </details>
+              <Markdown v-else :source="event.discussion.body" class="mt-[0.25em] text-sm" />
+              <div
+                v-if="event.discussion.reactions.length"
+                class="mt-[0.25em] flex flex-wrap gap-x-3"
+              >
+                <span
+                  v-for="reaction in event.discussion.reactions"
+                  :key="reaction.emoji"
+                  :title="`Reaction from ${new Intl.ListFormat('en', {
+                    style: 'long',
+                    type: 'conjunction',
+                  }).format(
+                    reaction.authors.map(
+                      (author) => resolveRadicleIdentity(author)?.alias ?? author,
+                    ),
+                  )}`"
+                  :class="{
+                    'modified-by-local-identity': reaction.authors.find((author) =>
+                      localIdentity?.id.includes(author),
+                    ),
+                  }"
+                >
+                  <template
+                    v-if="
+                      event.discussion.reactions.flatMap((reaction) => reaction.authors)
+                        .length <= 4
+                    "
+                  >
+                    {{ reaction.emoji }}
+                    <template
+                      v-for="(part, index) in new Intl.ListFormat('en', {
+                        style: 'short',
+                        type: 'unit',
+                      }).formatToParts(
+                        reaction.authors.map(
+                          (author) =>
+                            resolveRadicleIdentity(author)?.alias ?? truncateMiddle(author),
+                        ),
+                      )"
+                      :key="index"
+                    >
+                      <span v-if="part.type === 'element'" class="font-mono">{{
+                        part.value
+                      }}</span>
+                      <template v-else-if="part.type === 'literal'">{{ part.value }}</template>
+                    </template>
+                  </template>
+                  <template v-else>
+                    {{ reaction.emoji }}
+                    <span class="font-mono">{{ reaction.authors.length }}</span>
+                  </template>
+                </span>
+              </div>
+            </EventItem>
+            <EventItem
+              v-else-if="event.kind === 'merge'"
+              :when="event.ts"
+              codicon="codicon-git-merge"
+            >
+              Patch merged by
+              <span :title="event.merge.author.id" class="font-mono">{{
+                getIdentityAliasOrId(event.merge.author)
+              }}</span>
+              using revision
+              <span
+                v-if="event.revision"
+                @click="selectAndScrollToRevision(event.revision)"
+                :title="getRevisionHoverTitle(event.revision.description)"
+                class="font-mono hover:cursor-pointer"
+                >{{ shortenHash(event.revision.id) }}</span
+              >
+              <span v-else class="font-mono">{{ shortenHash(event.merge.revision) }}</span>
+            </EventItem>
+          </template>
+        </EventList>
       </section>
     </main>
   </article>
 </template>
 
 <style scoped>
-.timeline {
-  position: relative;
+.modified-by-local-identity {
+  @apply outline outline-1 outline-offset-1;
+  background-color: color-mix(in srgb, var(--vscode-editor-foreground) 7%, transparent);
+  outline-color: color-mix(in srgb, var(--vscode-editor-foreground) 20%, transparent);
+}
 
-  &:before {
-    content: '';
-    position: absolute;
-    z-index: -1;
-    left: 33.4px;
-    height: 100%;
-    width: 1.2px;
-    background-color: color-mix(
-      in srgb-linear,
-      var(--vscode-editor-background),
-      var(--vscode-editor-foreground) 5%
-    );
+:deep(.pulse-outline) {
+  @keyframes outline-pulse {
+    from {
+      @apply outline-transparent;
+    }
+    25% {
+      @apply outline-vscode-editor-foreground;
+    }
+    50% {
+      @apply outline-transparent;
+    }
+    70% {
+      @apply outline-vscode-editor-foreground;
+    }
+    to {
+      @apply outline-transparent;
+    }
   }
+
+  @apply outline outline-offset-[0.25em];
+  animation: outline-pulse 1.5s ease-in-out forwards;
 }
 </style>